shellserver/requirements.txt
mcp-python-sdk>=1.4.1 